+++ outputs.tf
@@ -42,3 +42,8 @@ output "api_management_scm_url" {
   description = "The URL for the SCM Endpoint associated with this API Management service"
   value       = azurerm_api_management.apim.scm_url
 }
+
+output "api_management_identity" {
+  description = "The identity map of the APIM"
+  value       = azurerm_api_management.apim.identity
+}