+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v3.1.0 - Unreleased
+
+Added
+  * AZ-264: Added new variable to handle identity plus corresponding output
+  * AZ-272: named values, products, groups and their relationships
+  
+Fixed
+  * AZ-358: Fix RPC Error when using empty hostname_configurations
+
 # v3.0.0 - 2020-07-09
 
 Breaking